A simple Profiler class.
Adapted the Profiler code from Kerbalism for use with KSPTrajectories.
Also some helper functions placed in Util.cs to support the profiler,
they are also handy for use in the rest of the code base.

The profiler is enabled/disabled with the compiler directive
TRAJECTORIES_PROFILER.

Use the Start and Stop methods to wrap the code you want to measure.

Start KSP, load a game and press Ctrl+p to open the profiler window.

The helper functions used by the Utilities/profiler.cs file, are from
Kerbalism's Lib.cs file. They include:
- Time methods
- Game Logic
- Random number generators.