Profiler: Entries can now be recorded by Telemetry.
Add PROFILER_TELEMETRY to the conditional compilation symbols to enable the Profiler telemetry.
Number of calls and execution time is recorded to the Telemetry file for every entry in the Profiler.
You will have to run the profiler at least once and then re-load the scene before the Telemetry module can get a list of the entries and create a valid Telemetry.csv file.